# gateway/api/swagger.yaml
basePath: /
definitions:
...
  httpx.ErrorResponse:
    properties:
      error:
        example: access denied
        type: string
    type: object
...
  report.CreateDailyReportCommentRequest:
    properties:
      comment:
        type: string
    type: object
...
  report.DailyReportCommentResponse:
    properties:
      author_user_id:
        type: string
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      report_id:
        type: string
    type: object
...
  report.ListDailyReportCommentsResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/report.DailyReportCommentResponse'
        type: array
    type: object
...
  report.UpdateDailyReportCommentRequest:
    properties:
      comment:
        type: string
    type: object
...
host: localhost:28080
info:
  contact: {}
  description: Public REST API exposed by the gateway. Internal microservices communicate
    only over gRPC and are intentionally omitted from this specification.
  title: Project Management Gateway API
  version: "1.0"
paths:
...
  /api/report-comments/{id}:
    delete:
      description: Deletes a daily report comment. Accessible to the comment author
        and administrators.
      parameters:
      - description: Daily report comment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete daily report comment
      tags:
      - report-comments
    get:
      description: Returns a daily report comment by ID. Accessible to users who can
        read the parent report.
      parameters:
      - description: Daily report comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/report.DailyReportCommentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get daily report comment
      tags:
      - report-comments
    put:
      consumes:
      - application/json
      description: Updates a daily report comment. Accessible to the comment author
        and administrators.
      parameters:
      - description: Daily report comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Daily report comment update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/report.UpdateDailyReportCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/report.DailyReportCommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update daily report comment
      tags:
      - report-comments
...
  /api/reports/{reportId}/comments:
    get:
      description: Returns comments attached to a daily report. Accessible to users
        who can read the report.
      parameters:
      - description: Daily report ID
        in: path
        name: reportId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/report.ListDailyReportCommentsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List daily report comments
      tags:
      - report-comments
    post:
      consumes:
      - application/json
      description: Adds a comment to a daily report. Accessible to users who can read
        the report, including the author, administrators, and reviewers.
      parameters:
      - description: Daily report ID
        in: path
        name: reportId
        required: true
        type: string
      - description: Daily report comment payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/report.CreateDailyReportCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/report.DailyReportCommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create daily report comment
      tags:
      - report-comments
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: JWT access token. Use the `Bearer {token}` format.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"